+++ .storybook/theme-selector/register.js
@@ -48,6 +48,15 @@ export const ThemeSwitcher = memo(
     const [activeTheme, setTheme] = useState(getThemeName());
     const [expanded, setExpanded] = useState(false);
 
+    if (process.env.STORYBOOK_DEBUG_ALL_THEMES) {
+      modernThemes.all = {
+        colors: {
+          primary:
+            "linear-gradient(45deg, red, orange, yellow, green, blue, indigo, violet, red)",
+        },
+      };
+    }
+
     const themeList = Object.keys(modernThemes).map((themeName) => ({
       id: themeName,
       title: themeName,