+++ cypress/support/step-definitions/help-steps.js
@@ -1,5 +1,10 @@
 import helpHref from "../../locators/help";
+import icon from "../../locators";
 
 Then("Help href on preview is set to {word}", (href) => {
   helpHref().should("have.attr", "href", href);
+});
+
+Then("Help component is focused", () => {
+  icon().should("be.focused");
 });