+++ src/__internal__/tooltip-provider/index.js
@@ -3,8 +3,12 @@ import PropTypes from "prop-types";
 
 export const TooltipContext = React.createContext({});
 
-export const TooltipProvider = ({ children, tooltipPosition }) => (
-  <TooltipContext.Provider value={{ tooltipPosition }}>
+export const TooltipProvider = ({
+  children,
+  tooltipPosition,
+  helpAriaLabel,
+}) => (
+  <TooltipContext.Provider value={{ tooltipPosition, helpAriaLabel }}>
     {children}
   </TooltipContext.Provider>
 );
@@ -12,4 +16,5 @@ export const TooltipProvider = ({ children, tooltipPosition }) => (
 TooltipProvider.propTypes = {
   children: PropTypes.node.isRequired,
   tooltipPosition: PropTypes.oneOf(["top", "bottom", "left", "right"]),
+  helpAriaLabel: PropTypes.string,
 };