+++ src/components/icon/icon.spec.js
@@ -140,7 +140,7 @@ describe("Icon component", () => {
       "#123456",
     ];
     describe.each(correctColors)("when color prop is provided", (color) => {
-      it("takes precedence over iconColor and renders properly colored Icon", () => {
+      it("renders properly colored Icon", () => {
         const wrapper = mount(<Icon type="home" color={color} />);
         const { color: renderedColor } = styledColor({
           theme: baseTheme,
@@ -168,7 +168,7 @@ describe("Icon component", () => {
         );
       });
 
-      it("takes precedence over iconColor and renders properly colored Icon with tooltip", () => {
+      it("renders properly colored Icon with tooltip", () => {
         const wrapper = mount(
           <Icon
             type="home"
@@ -219,7 +219,7 @@ describe("Icon component", () => {
       });
     });
     describe.each(correctColors)("when bg prop is provided", (color) => {
-      it("takes precedence over bgTheme and renders properly colored Icon", () => {
+      it("renders properly colored Icon", () => {
         const wrapper = mount(<Icon bg={color} type="message" />);
         const { backgroundColor } = styledColor({
           theme: baseTheme,
@@ -289,130 +289,13 @@ describe("Icon component", () => {
 
   describe("icon color", () => {
     it("renders proper icon color for disabled state", () => {
-      const wrapper = renderStyles({ disabled: true, bgTheme: "business" });
+      const wrapper = renderStyles({ disabled: true });
       assertStyleMatch(
         {
           color: baseTheme.icon.disabled,
         },
         wrapper.toJSON()
       );
-    });
-
-    describe.each(["error", "info", "business"])(
-      "when the background theme is %s",
-      (whiteIconBackground) => {
-        it("renders a white icon", () => {
-          const wrapper = renderStyles({ bgTheme: whiteIconBackground });
-          assertStyleMatch(
-            {
-              color: baseTheme.colors.white,
-            },
-            wrapper.toJSON()
-          );
-        });
-      }
-    );
-
-    describe.each(["success", "warning"])(
-      "when the background theme is %s",
-      (darkIconBackground) => {
-        it("renders a dark icon", () => {
-          const wrapper = renderStyles({
-            bgTheme: darkIconBackground,
-            isInteractive: true,
-          });
-          assertStyleMatch(
-            {
-              color: baseTheme.icon.default,
-            },
-            wrapper.toJSON()
-          );
-
-          assertStyleMatch(
-            {
-              color: baseTheme.icon.defaultHover,
-            },
-            wrapper.toJSON(),
-            { modifier: ":hover" }
-          );
-        });
-      }
-    );
-
-    describe("when bgTheme is set to none", () => {
-      it("renders white icon when iconColor is set to onDarkBackground", () => {
-        const wrapper = renderStyles({
-          iconColor: "on-dark-background",
-          bgTheme: "none",
-        });
-        assertStyleMatch(
-          {
-            color: baseTheme.colors.white,
-          },
-          wrapper.toJSON()
-        );
-      });
-
-      it("renders dark icon when iconColor is set to onLightBackground", () => {
-        const wrapper = renderStyles({
-          iconColor: "on-light-background",
-          bgTheme: "none",
-          isInteractive: true,
-        });
-        assertStyleMatch(
-          {
-            color: baseTheme.icon.onLightBackground,
-          },
-          wrapper.toJSON()
-        );
-
-        assertStyleMatch(
-          {
-            color: baseTheme.icon.onLightBackgroundHover,
-          },
-          wrapper.toJSON(),
-          { modifier: ":hover" }
-        );
-      });
-
-      describe('fontSize is "large" and bgSize is "small"', () => {
-        it('sets the height and width style properties of bgSize to same as "large"', () => {
-          const wrapper = renderStyles({
-            fontSize: "large",
-            bgTheme: "business",
-            bgSize: "small",
-          });
-          assertStyleMatch(
-            {
-              height: "40px",
-              width: "40px",
-            },
-            wrapper.toJSON()
-          );
-        });
-      });
-
-      it("renders dark icon when iconColor is set to business-color", () => {
-        const wrapper = renderStyles({
-          iconColor: "business-color",
-          bgTheme: "none",
-          isInteractive: true,
-        });
-        assertStyleMatch(
-          {
-            color: baseTheme.colors.primary,
-          },
-          wrapper.toJSON()
-        );
-
-        assertStyleMatch(
-          {
-            color: "#006800",
-          },
-          wrapper.toJSON(),
-          { modifier: ":hover" }
-        );
-      });
     });
   });
 
@@ -425,72 +308,6 @@ describe("Icon component", () => {
             backgroundColor: baseTheme.icon.disabled,
           },
           wrapper.toJSON()
-        );
-      });
-    });
-
-    describe("when bgTheme is set to none", () => {
-      it("renders transparent background", () => {
-        const wrapper = renderStyles({ bgTheme: "none" });
-        assertStyleMatch(
-          {
-            backgroundColor: "transparent",
-          },
-          wrapper.toJSON()
-        );
-      });
-    });
-
-    describe.each(["info", "error", "success", "warning"])(
-      "when bgTheme is set to one of the statuses",
-      (status) => {
-        const wrapper = renderStyles({ bgTheme: status, isInteractive: true });
-        const hoverColors = {
-          info: "#005C9B",
-          error: "#9F2D3F",
-          success: "#008D00",
-          warning: "#BA5000",
-        };
-
-        it(`renders proper background color for ${status}`, () => {
-          assertStyleMatch(
-            {
-              backgroundColor: baseTheme.colors[status],
-            },
-            wrapper.toJSON()
-          );
-
-          assertStyleMatch(
-            {
-              backgroundColor: hoverColors[status],
-            },
-            wrapper.toJSON(),
-            { modifier: ":hover" }
-          );
-        });
-      }
-    );
-
-    describe("when bgTheme is set to business", () => {
-      const wrapper = renderStyles({
-        bgTheme: "business",
-        isInteractive: true,
-      });
-
-      it("renders proper background color", () => {
-        assertStyleMatch(
-          {
-            backgroundColor: baseTheme.colors.primary,
-          },
-          wrapper.toJSON()
-        );
-
-        assertStyleMatch(
-          {
-            backgroundColor: "#006800",
-          },
-          wrapper.toJSON(),
-          { modifier: ":hover" }
         );
       });
     });
@@ -589,6 +406,54 @@ describe("Icon component", () => {
       );
 
       expect(wrapper.find(Tooltip).props().isVisible).toEqual(false);
+    });
+
+    it("sets the tabIndex, ariaLabel and role", () => {
+      const wrapper = mount(
+        <Icon
+          type="home"
+          ariaLabel="home"
+          tooltipMessage="foo"
+          tooltipVisible
+        />
+      );
+
+      expect(wrapper.find(StyledIcon).props()).toEqual(
+        expect.objectContaining({
+          "aria-label": "home",
+          role: "tooltip",
+          tabIndex: 0,
+        })
+      );
+    });
+
+    it("does not set the tabIndex and role when the disabled prop is set", () => {
+      const wrapper = mount(
+        <Icon type="home" tooltipMessage="foo" tooltipVisible disabled />
+      );
+
+      expect(wrapper.find(StyledIcon).props()).toEqual(
+        expect.objectContaining({
+          role: undefined,
+          tabIndex: undefined,
+        })
+      );
+    });
+
+    it("supports overriding the tabIndex and role", () => {
+      const wrapper = mount(
+        <Icon
+          type="home"
+          ariaLabel="home"
+          tooltipMessage="foo"
+          tooltipVisible
+          role="region"
+          tabIndex={-1}
+        />
+      );
+
+      expect(wrapper.find(StyledIcon).prop("role")).toEqual("region");
+      expect(wrapper.find(StyledIcon).prop("tabIndex")).toEqual(-1);
     });
 
     describe("tooltipFlipOverrides", () => {