+++ src/components/icon/icon.style.js
@@ -5,67 +5,11 @@ import { margin } from "styled-system";
 
 import iconUnicodes from "./icon-unicodes";
 import baseTheme from "../../style/themes/base";
-import generatePalette from "../../style/palette";
-import iconSizeConfig, {
-  ICON_BACKGROUNDS,
-  ICON_COLORS,
-  ICON_SHAPES,
-} from "./icon-config";
+import iconSizeConfig, { ICON_SHAPES } from "./icon-config";
 import browserTypeCheck, {
   isSafari,
 } from "../../__internal__/utils/helpers/browser-type-check";
 import styledColor from "../../style/utils/color";
-
-const getBackgroundColor = (theme, bgTheme, disabled, isHover) => {
-  if (bgTheme !== "none") {
-    if (disabled) return theme.icon.disabled;
-  }
-
-  const palette = generatePalette({
-    statusColor: theme.colors[bgTheme],
-    businessColor: theme.colors.primary,
-  });
-  const statuses = ["info", "error", "success", "warning"];
-  if (statuses.includes(bgTheme)) {
-    return isHover ? palette.statusColorShade(20) : theme.colors[bgTheme];
-  }
-
-  if (bgTheme === "business") {
-    return isHover ? palette.businessColorShade(20) : theme.colors.primary;
-  }
-
-  return "transparent";
-};
-
-const getIconColor = (bgTheme, theme, iconColor, disabled, isHover) => {
-  if (disabled) return theme.icon.disabled;
-
-  if (bgTheme !== "none") {
-    const whiteIconBackgrounds = ["error", "info", "business"];
-    const darkIconBackgrounds = ["success", "warning"];
-
-    if (whiteIconBackgrounds.includes(bgTheme)) return theme.colors.white;
-
-    if (darkIconBackgrounds.includes(bgTheme)) {
-      return isHover ? theme.icon.defaultHover : theme.icon.default;
-    }
-  }
-
-  const palette = generatePalette({ businessColor: theme.colors.primary });
-
-  switch (iconColor) {
-    case "on-dark-background":
-      return theme.colors.white;
-    case "on-light-background":
-      return isHover
-        ? theme.icon.onLightBackgroundHover
-        : theme.icon.onLightBackground;
-    case "business-color":
-      return isHover ? palette.businessColorShade(20) : theme.colors.primary;
-    default:
-      return isHover ? theme.icon.defaultHover : theme.icon.default;
-  }
-};
 
 function adjustIconBgSize(fontSize, bgSize) {
   const replacements = {
@@ -94,17 +38,16 @@ function adjustIconBgSize(fontSize, bgSize) {
 
 const StyledIcon = styled.span`
   ${({
-    bgTheme,
     theme,
     color,
     bg,
     isInteractive,
-    iconColor,
     bgSize,
     bgShape,
     type,
     fontSize,
     disabled,
+    hasTooltip,
   }) => {
     let finalColor;
     let finalHoverColor;
@@ -112,40 +55,44 @@ const StyledIcon = styled.span`
     let bgHoverColor;
 
     try {
-      if (color) {
+      if (disabled) {
+        finalColor = theme.icon.disabled;
+        finalHoverColor = theme.icon.disabled;
+      } else if (color) {
         const { color: renderedColor } = styledColor({ color, theme });
         finalColor = renderedColor;
         finalHoverColor = shade(0.2, renderedColor);
       } else {
-        finalColor = getIconColor(bgTheme, theme, iconColor, disabled, false);
-        finalHoverColor = getIconColor(
-          bgTheme,
-          theme,
-          iconColor,
-          disabled,
-          true
-        );
+        finalColor = theme.icon.default;
+        finalHoverColor = theme.icon.defaultHover;
       }
 
       if (bg) {
         const { backgroundColor } = styledColor({ bg, theme });
         bgColor = backgroundColor;
         bgHoverColor = shade(0.2, backgroundColor);
+      } else if (disabled) {
+        bgColor = theme.icon.disabled;
+        bgHoverColor = theme.icon.disabled;
       } else {
-        bgColor = getBackgroundColor(theme, bgTheme, disabled, false);
-        bgHoverColor = getBackgroundColor(theme, bgTheme, disabled, true);
+        bgColor = "transparent";
+        bgHoverColor = "transparent";
       }
     } catch (e) {
       // eslint-disable-next-line no-console
       console.error(e);
     }
-
     return css`
-      display: inline-block;
       position: relative;
       color: ${finalColor};
       background-color: ${bgColor};
       vertical-align: middle;
+      align-items: center;
+      display: inline-flex;
+      justify-content: center;
+      height: ${adjustIconBgSize(fontSize, bgSize)};
+      width: ${adjustIconBgSize(fontSize, bgSize)};
+      border-radius: ${iconSizeConfig.backgroundShape[bgShape]};
 
       ${isInteractive &&
       css`
@@ -153,16 +100,6 @@ const StyledIcon = styled.span`
           color: ${finalHoverColor};
           background-color: ${bgHoverColor};
         }
-      `}
-
-      ${bgTheme !== "none" &&
-      css`
-        align-items: center;
-        display: inline-flex;
-        justify-content: center;
-        height: ${adjustIconBgSize(fontSize, bgSize)};
-        width: ${adjustIconBgSize(fontSize, bgSize)};
-        border-radius: ${iconSizeConfig.backgroundShape[bgShape]};
       `}
 
       &::before {
@@ -193,6 +130,13 @@ const StyledIcon = styled.span`
         display: block;
       }
 
+      ${hasTooltip &&
+      `
+        :focus {
+          outline: 2px solid ${theme.colors.focus};
+        }
+      `}
+
       ${margin}
     `;
   }}
@@ -211,9 +155,7 @@ StyledIcon.propTypes = {
     "extra-large",
   ]),
   bgShape: PropTypes.oneOf(ICON_SHAPES),
-  bgTheme: PropTypes.oneOf([...ICON_COLORS, ...ICON_BACKGROUNDS, ""]),
   fontSize: PropTypes.oneOf(["small", "medium", "large", "extra-large"]),
-  iconColor: PropTypes.oneOf(ICON_COLORS),
 };
 
 StyledIcon.defaultProps = {